favorite files are ready to handle the favorite data from the favorite API
