+++ src/pages/Home.js
@@ -8,7 +8,7 @@ import favoritesList from '../utils/get-favorites.js';
 const Home = (API, USER) => {
 
   /* crear una condicion para verificar si existe una llave de favoritos en el localstorage */
-
+/*
   if (favoritesList(API)) {
     favoritesList(API)
       .then(result => {
@@ -19,7 +19,7 @@ const Home = (API, USER) => {
       });
   } else {
     console.log('user without favorite data');
-  }  
+  }   */
   
   Header(USER);
   Hero();
